pyproject.toml
[tool.poetry]
name = "nlp-ad"
version = "0.1.0"
description = ""
authors = ["Scott Bailey <csb5t@virginia.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
ipykernel = "^5.4.3"
nltk = "^3.5"
matplotlib = "^3.3.4"
jupyterlab = "^3.0.7"
textacy = "^0.11.0"
streamlit = "^0.81.0"
spacy-streamlit = "^1.0.0"
transformers = "^4.5.1"
seaborn = "^0.11.1"
altair = "^4.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
mypy = "^0.812"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
